Refactor align rule
